Fixed bug in trans method of coord class
Recent changes to the roll function where the rolled coord of   (e.g.
xt) a field’s grid is replaced with a version (e.g. xt_rolled)
reflecting the roll led to a bug in the trans method. Namely, a
translated field could no longer be subtracted easily from the original
field, throwing an error. As a result, for instance, stream functions
could no longer be computed as the derivative of the velocity could no
longer be taken. This has now been fixed by allowing a new argument
flag “keepgrid” that is False by default, but can be set to True in
trans, coord_shift and roll to keep the original grid of the field (and
not replace to coord as per above).